Hopefully last set of changes for 1.0 release
Added save data entry for 3D scenes state, so if you last used it with the 3D state on, when you reopen the application it will auto start the 3D scene.

Went through all of the functions and made them static typed(Just to be consistent)

Added desktop icon.

Bumped up version from 1.0rc to 1.0